1. 转换.net7 2. 优化 3. 修复大数字转换为.0的问题

+++ JsonExporter.cs
@@ -148,7 +148,7 @@ namespace excel2json
                             object cellJsonObj = JsonConvert.DeserializeObject(cellText);
                             if (cellJsonObj != null)
                                 value = cellJsonObj;
-                        }
+                        }                        
                         catch (Exception exp)
                         {
                         }
@@ -162,8 +162,8 @@ namespace excel2json
                 else if (value.GetType() == typeof(double))
                 { // 去掉数值字段的“.0”
                     double num = (double)value;
-                    if ((int)num == num)
-                        value = (int)num;
+                    if ((long)num == num)
+                        value = (long)num;
                 }
 
                 //全部转换为string
